Add data preparation and content generation modules with article processing capabilities remove unused artefacts
