+++ createArticles/dataPrep/__init__.py
@@ -1,0 +1,18 @@
+from createArticles.dataPrep.similarity import (
+    cosine_similarity, 
+    check_for_similar_articles,
+    check_processed_articles_similarity
+)
+
+from createArticles.dataPrep.article_grouping import (
+    group_similar_articles,
+    process_article_group
+)
+
+__all__ = [
+    'cosine_similarity',
+    'check_for_similar_articles',
+    'check_processed_articles_similarity',
+    'group_similar_articles',
+    'process_article_group'
+]
