+++ createArticles/prompts.yaml
@@ -32,61 +32,6 @@ german_prompt: |
      - Die Überschrift muss als HTML mit dem <h1>-Tag formatiert werden.
      - Der Artikelinhalt muss als HTML mit geeigneten Tags formatiert werden, wie z. B. <p> für Absätze, <h2> oder <h3> für Unterüberschriften und <div> als Container.
 
-post_processing_prompt: |
-  You are an experienced editor specialized in digital publishing, expert in content curation and SEO optimization.
-  Review the following article content to ensure it is perfectly curated, content-rich, and optimized for search engines. Perform the following checks and enhancements:
-
-  **Curation and Content Enhancement:**
-  1. **Fact Verification & Accuracy:**
-     - Carefully review the article content and cross-reference key facts and figures against reliable sources (if needed based on your knowledge of the topic). Correct any factual inaccuracies or misinterpretations.
-  2. **Clarity, Flow & Narrative:**
-     - Evaluate the article's clarity and logical flow. Ensure smooth transitions between paragraphs and ideas. Rewrite any confusing sentences or paragraphs for better readability and coherence.
-     - Assess the overall narrative. Is the story engaging and easy to follow? Enhance the narrative flow if needed to create a more compelling reading experience.
-  3. **Engagement & Tone Polish:**
-     - Review the article's tone and ensure it aligns with the requested dynamic, energetic, and engaging style.
-     - Enhance the language by incorporating even more vivid adjectives, stronger verbs, and descriptive imagery.
-     - Identify opportunities to strategically insert rhetorical questions or engaging sentence structures to actively involve the reader and maintain their interest.
-  4. **Relevance & Reader Interest:**
-     - Consider the target audience (sports news readers). Does the article effectively capture their interest and provide relevant information?
-     - Identify if any aspects could be expanded upon or clarified to increase reader understanding and engagement. Suggest adding brief contextual details if necessary to enhance relevance.
-  5. **Completeness & Depth:**
-     - Ensure the article adequately covers the main topic derived from the source material.
-     - Check if all key points from the "main_content" are effectively presented in the "Body" section.
-     - Evaluate if the conclusion effectively summarizes the article and provides a satisfying closing.
-
-  **SEO Optimization:**
-  6. **Keyword Optimization (Headline & Body):**
-     - Analyze the headline and body content for relevant sports-related keywords.
-     - Verify if keywords are naturally integrated within the text, especially in the headline, subheadings, and the first paragraph.
-     - Suggest incorporating additional relevant keywords (without keyword stuffing) where contextually appropriate to improve search engine visibility for target queries.
-  7. **Headline SEO Effectiveness:**
-     - Critically assess the headline's SEO effectiveness. Is it concise, compelling, and keyword-rich?
-     - Ensure the headline accurately reflects the article's content and is optimized to attract clicks from search engine results pages. Suggest improvements if needed to enhance click-through rate and SEO value.
-  8. **Subheading Optimization (SEO & Readability):**
-     - Review subheadings (if present) for both SEO value and readability.
-     - Ensure subheadings are informative, keyword-relevant, and effectively break down the text for better readability.
-     - Suggest improvements to subheadings to enhance their SEO impact and guide readers through the article's structure.
-
-  **Editorial Polish & Formatting:**
-  9. **Grammar, Spelling & Punctuation Perfection:**
-     - Conduct a meticulous grammar and spell check to eliminate any errors.
-     - Ensure perfect punctuation throughout the article for professional presentation.
-  10. **Style Consistency & Journalistic Standards:**
-      - Verify consistency in tone, voice, and writing style across the entire article.
-      - Confirm that the article adheres to general journalistic writing standards for clarity, objectivity, and engaging storytelling.
-  11. **Conciseness & Word Count Adherence:**
-      - Check if the article falls within the desired word count range (350-500 words).
-      - Trim any unnecessary words, phrases, or redundant sentences to improve conciseness without sacrificing essential content.
-  12. **HTML Validity & Cleanliness:**
-      - Re-examine the HTML code for validity and correctness.
-      - Remove any extraneous characters, unwanted formatting artifacts, or broken HTML tags.
-      - Ensure all HTML tags are correctly opened and closed and that the structure adheres to the requested format: `<h1>` for headline, `<p>`, `<h2>`/`<h3>` for content and subheadings, and `<div>` as container.
-
-  **Output:** Return only the thoroughly reviewed and enhanced article text in clean and valid HTML format, ready for digital publishing, ensuring perfect curation, content excellence and SEO optimization.
-
-  Article content:
-  {article_content}
-
 team_detection_prompt: |
   You are an expert in sports news analysis. Analyze the following article content and determine which team is the primary subject. Also assign a confidence score between 0 and 1 to your detection. Return the answer in JSON format with exactly two keys:
   - "team": a string containing the team's name.
